Update Email.cs
More flexibility by sending a separate model. Useful when preparing a model and then send it to an external service like Hangfire.
The preparation of the model is done by the program, while the use is delegated to another service that must be able to reconstruct the model with data. I tried using the object Posta.Email, but the serialization fails for its complexity and class inheritance from DynamicObject.