Change access to member ImageEmbedder on Email object from internal to public
We need to extend the existing functionality of the EmailParser with
different handling of the Postal Header and integrate functionality to
move css styles inline. Therefore we need to write our own EmailParser
implementing IEmailParser Interface. Unfortunately with the internal on
the Email object we can't use the ImageEmbedder in our Parser. To ensure
extensibility of the EmailParser the ImageEmbedder should be accessable.