Change EmailViewResult to use the final Parser output, rather than the result output.
EmailViewResult now gives an accurate output if a custom parser is used (e.g. when inlining the CSS), as it now uses the MailMessage body as the output, rather than the Renderer's result.

Now also checks for alternative views if no format query is given, to render emails with embedded images.